fixing bug: changing input type from 'int' to 'int64' to allow reading in also the MODE_OF_OPERATION and add it to typed_index

+++ workflow/scripts/create_modelrun.py
@@ -94,7 +94,7 @@ def get_types_from_tuple(index: list, param: str, config: Dict) -> Tuple:
             typed_index.append(str(element))
         elif this_type == 'float':
             typed_index.append(float(element))
-        elif this_type == 'int':
+        elif this_type == 'int64':
             typed_index.append(int(element))
 
     return typed_index
@@ -224,7 +224,6 @@ def apply_interploted_values(
     ValueError
         If index not found in original dataframe 
     """
-
     try:
         model_params[name].loc[tuple(index + [first_year]):tuple(index + [end_year]), 'VALUE'] = interpolated_values[0]
     except ValueError as ex:
